+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Added support for custom handling of settings updates.
 
 ### Changed
+* [breaking] MqttClient constructor now accepts initial settings values.
+
 ### Removed
 
 ## [0.3.0] - 2021-12-13